+++ __tests__/index.test.js
@@ -1,14 +1,22 @@
-import fs from "fs";
-import path from "path";
-import eslint from "eslint";
+import * as fs from "fs";
+import * as path from "path";
+import * as eslint from "eslint";
 import test from "ava";
 import { configs } from "../index.js";
 import { devDependencies, peerDependencies } from "../package.json";
 
+/**
+ * @param {any} obj Maybe object
+ * @returns {boolean} Is object?
+ */
 function isObject(obj) {
   return typeof obj === "object" && obj !== null;
 }
 
+/**
+ * @param {string} input Any input
+ * @returns {string} String transformed from dash to camel case
+ */
 function dash2CamelCase(input) {
   return input.replace(/-([a-z])/gu, (found) => found[1].toUpperCase());
 }
@@ -27,6 +35,7 @@ test("should the `eslint` and the `eslint` CLI present", (t) => {
 
 test("should all configs are present in exports", (t) => {
   const configDir = path.resolve(__dirname, "../lib/config");
+  /** @type {string[]} */
   let files = [];
 
   // eslint-disable-next-line node/no-sync
@@ -55,7 +64,7 @@ test("should load the 'base' preset", (t) => {
 
   // t.is(configForFile.parser, require.resolve("babel-eslint"));
   t.deepEqual(configForFile.parserOptions, {
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
 
@@ -98,7 +107,7 @@ test("should load the 'script' preset", (t) => {
     ecmaFeatures: {
       globalReturn: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.true(configForFile.plugins.includes("import"));
@@ -141,7 +150,7 @@ test("should load the 'module' preset", (t) => {
   t.is(configForFile.parser, require.resolve("babel-eslint"));
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
 
@@ -176,7 +185,7 @@ test("should load the 'dirty' preset", (t) => {
   t.is(configForFile.parser, require.resolve("babel-eslint"));
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
     ecmaFeatures: { globalReturn: true },
   });
@@ -211,7 +220,7 @@ test("should load the 'ava' preset", (t) => {
 
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
 
@@ -245,7 +254,7 @@ test("should load the 'esnext' preset", (t) => {
 
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
   t.true(configForFile.env.es2020);
@@ -277,7 +286,7 @@ test("should load the 'lodash' preset", (t) => {
 
   t.deepEqual(configForFile.parserOptions, {
     allowImportExportEverywhere: true,
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
 
@@ -313,7 +322,7 @@ test("should load the 'node' preset", (t) => {
     ecmaFeatures: {
       globalReturn: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.true(configForFile.env.node);
@@ -348,7 +357,7 @@ test("should load the 'browser' preset", (t) => {
     ecmaFeatures: {
       globalReturn: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.false(Boolean(configForFile.env.node));
@@ -388,7 +397,7 @@ test("should load 'node' and 'browser' presets", (t) => {
     ecmaFeatures: {
       globalReturn: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.true(configForFileNodeAndBrowser.env.node);
@@ -441,7 +450,7 @@ test("should load 'node' and 'browser' presets", (t) => {
     ecmaFeatures: {
       globalReturn: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.true(configForFileBrowserAndNode.env.node);
@@ -492,7 +501,7 @@ test("should load 'node' and 'browser' presets", (t) => {
     ecmaFeatures: {
       globalReturn: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.true(configForFileESNextLast.env.node);
@@ -541,7 +550,7 @@ test("should load the 'react' preset", (t) => {
     ecmaFeatures: {
       jsx: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
   t.true(configForFile.env.browser);
@@ -635,6 +644,27 @@ test("should load the 'jest' preset", (t) => {
   t.is(report.warningCount, 0, "eslint report without warnings");
 });
 
+test("should load the 'jsdoc-typescript' preset", (t) => {
+  const cli = new eslint.CLIEngine({
+    useEslintrc: false,
+    baseConfig: {
+      extends: ["./lib/config/jsdoc-typescript", "./lib/config/module"],
+    },
+  });
+
+  const configForFile = cli.getConfigForFile("myfile.js");
+
+  t.true(configForFile.plugins.includes("jsdoc"));
+
+  const report = cli.executeOnFiles([
+    path.resolve(__dirname, "./fixtures/jsdoc-typescript.js"),
+  ]);
+
+  t.is(report.results.length, 1, "eslint report with one results");
+  t.is(report.errorCount, 0, "eslint report without errors");
+  t.is(report.warningCount, 0, "eslint report without warnings");
+});
+
 test("should load the 'markdown' preset", (t) => {
   const cli = new eslint.CLIEngine({
     useEslintrc: false,
@@ -698,7 +728,7 @@ alert("test");
       globalReturn: true,
       impliedStrict: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "script",
   });
   t.true(scriptConfigForFile.plugins.includes("markdown"));
@@ -759,7 +789,7 @@ console.log(test);
     ecmaFeatures: {
       impliedStrict: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
   t.true(moduleConfigForFile.plugins.includes("markdown"));
@@ -819,7 +849,7 @@ console.log(test);
       globalReturn: true,
       impliedStrict: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
   t.true(dirtyConfigForFile.plugins.includes("markdown"));
@@ -894,6 +924,7 @@ test("should load the 'all' preset", (t) => {
     useEslintrc: false,
     baseConfig: {
       extends: [
+        "./lib/config/jsdoc-typescript.js",
         "./lib/config/dirty.js",
         "./lib/config/node.js",
         "./lib/config/browser.js",
@@ -917,9 +948,10 @@ test("should load the 'all' preset", (t) => {
       impliedStrict: true,
       jsx: true,
     },
-    ecmaVersion: 2020,
+    ecmaVersion: 2021,
     sourceType: "module",
   });
+  t.true(configForFile.plugins.includes("jsdoc"));
   t.true(configForFile.plugins.includes("ava"));
   t.true(configForFile.plugins.includes("import"));
   t.true(configForFile.plugins.includes("jsx-a11y"));
@@ -933,7 +965,16 @@ test("should load the 'all' preset", (t) => {
   t.true(configForFile.plugins.includes("react"));
 
   const report = cli.executeOnText(
-    "const value = 100;\n\nfunction foo() {\n  return 1;\n}\n\nfoo(value);\n",
+    `const value = 100;
+
+/**
+ * @returns {number} Number
+ */
+function foo() {
+  return 1;
+}
+
+foo(value);`,
     "test.js"
   );
 
@@ -945,7 +986,7 @@ test("should load the 'all' preset", (t) => {
 test("peerDependencies should be equal devDependencies", (t) => {
   for (const key in peerDependencies) {
     if ({}.hasOwnProperty.call(peerDependencies, key)) {
-      t.true(peerDependencies[key] === devDependencies[key]);
+      t.true(peerDependencies[key] === devDependencies[key], key);
     }
   }
 });